Few changes
Added a constructor to the App class where we can now store the API data inside of state. Added 'Axios' to the project so we can make a get request to the URL endpoint specified in the "server.js" file. For now we are just logging business results to the console.